add visit entity, repository, service and controller

+++ src/dto/request/login_request.go
@@ -3,5 +3,5 @@ package request
 type LoginRequest struct {
 	Identity string `json:"identity" validate:"required"`
 	Password string `json:"password" validate:"required,gt=0"`
-	RoleID   uint   `json:"role_id" validate:"required"`
+	RoleID   uint64 `json:"role_id" validate:"required"`
 }
