+++ src/dto/request/pet_request.go
@@ -11,5 +11,5 @@ type PetRequest struct {
 	Weight      float32 `json:"weight" validate:"required"`
 	UserID      uint64  `json:"user_id" validate:"required"`
 	Location    string  `json:"location" validate:"required"`
-	PetTypeId   uint    `json:"pet_type_id" validate:"required"`
+	PetTypeId   uint64  `json:"pet_type_id" validate:"required"`
 }
