+++ src/dto/request/update-pet.go
@@ -10,5 +10,5 @@ type UpdatePetRequest struct {
 	Color       string  `json:"color"`
 	Weight      float32 `json:"weight"`
 	Location    string  `json:"location"`
-	PetTypeId   uint    `json:"pet_type_id"`
+	PetTypeId   uint64  `json:"pet_type_id"`
 }
