+++ src/models/pet-type_model.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type PetType struct {
 	gorm.Model
-	ID   uint   `gorm:"primaryKey;autoIncrement"`
+	ID   uint64 `gorm:"primaryKey;autoIncrement"`
 	Name string `gorm:"unique;not null"`
 }
