+++ src/services/pet-image_service.go
@@ -8,7 +8,7 @@ import (
 	"pet_api/src/repositories"
 )
 
-func CreatePetImages(petID uint, form *multipart.Form) ([]models.PetImage, error) {
+func CreatePetImages(petID uint64, form *multipart.Form) ([]models.PetImage, error) {
 	images, err := helpers.ImagesToBase64(form)
 	if err != nil {
 		return nil, err
